Move replay route package
